+++ src/main/java/com/SySTomateAlgo/TomateAlgo/Controllers/ClientController.java
@@ -4,6 +4,7 @@ import com.SySTomateAlgo.TomateAlgo.Entities.Client;
 import com.SySTomateAlgo.TomateAlgo.Repositories.ClientRepository;
 import com.SySTomateAlgo.TomateAlgo.Services.ClientService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -30,7 +31,12 @@ public class ClientController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Client> findByEmail(@PathVariable Long id){
+    public ResponseEntity<?> findById(@PathVariable Long id){
+        Optional<Client> client = service.findById(id);
+        if (client.isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Cliente no encontrado");
+        }
+
         return service.findById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -38,12 +44,18 @@ public class ClientController {
 
     @PutMapping("/{id}")
     public ResponseEntity<Client> update(@PathVariable Long id, @RequestBody Client client){
+
         return ResponseEntity.ok(service.update(id, client));
     }
 
 
     @DeleteMapping("/{id}")
     public ResponseEntity<String> delete(@PathVariable Long id){
+        Optional<Client> client = service.findById(id);
+        if (client.isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Cliente no encontrado");
+        }
+
         service.delete(id);
         return ResponseEntity.ok("El cliente fue eliminado con exito");
     }
